cars-manager/cars/vehicle/serializers.py
from rest_framework import serializers

from cars.ad.models import CarModel


class CarModelSerializer(serializers.ModelSerializer):
    class Meta:
        model = CarModel
        fields = '__all__'
