cars-manager/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.30.0
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
django-admin-honeypot-updated-2021==1.2.0
django-allauth==0.51.0
django-cloudinary-storage==0.3.0
django-extra-views==0.14.0
django-ipware==4.0.2
django-storages==1.13.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
numpy==1.23.5
oauthlib==3.2.2
Pillow==9.3.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
